move coastal files
