+++ inductiva/coastal/__init__.py
@@ -1,4 +1,5 @@
 #pylint: disable=missing-module-docstring
+from . import simulators
 from .bathymetry import Bathymetry
 from .coastal_area import CoastalArea
 from .output import CoastalAreaOutput
