+++ inductiva/fluids/scenarios/__init__.py
@@ -11,5 +11,4 @@ from .fluid_tank import (
     CircularTankInlet,
 )
 from .heat_sink import HeatSink
-from .coastal_area import (Bathymetry, CoastalArea, CoastalAreaOutput)
 from ._post_processing import SPHSimulationOutput
