+++ inductiva/fluids/simulators/__init__.py
@@ -1,6 +1,4 @@
 #pylint: disable=missing-module-docstring
 from .splishsplash import SPlisHSPlasH
 from .dualsphysics import DualSPHysics
-from .xbeach import XBeach
-from .swash import SWASH
 from .openfoam import OpenFOAM
